fix(feld): three field broadcasts no listener could read, and the types that hid them
The blocking mypy gate had 22 findings, all of them in this cycle's own
code and none of them ever run: ci.yml triggers on main only, so a
develop branch that is 83 commits long had never seen it.

One was a real defect. `broadcast_incident_update` takes the incident as
a dict; «Neue Meldung», its correction, and the Übung's simulated field
report each passed the bare id STRING, under the invented actions
«created» / «updated». Nothing could have read that as an incident — it
stayed invisible only because the board's listener ignores the body and
reloads. They now send {"id", "status"} under create / update, like
every other sender in the app.

The rest was one annotation lying, plus small stuff around it.
`Incident.location_lat/lng` said `float` while the column is Numeric, so
SQLAlchemy hands back a Decimal: correcting the two annotations removed
four findings at their source and introduced none. `.tuples()` gives the
row-to-dict conversions a real type; the four ignores that had stopped
being needed are gone (function types are a lookup table now, so both
sides are plain str); the DML count goes through `execute_dml`, which
exists for exactly this.

mypy: 0 in 133 files. 2860 backend tests green.

+++ backend/app/models.py
@@ -2,6 +2,7 @@
 
 import secrets
 from datetime import datetime
+from decimal import Decimal
 from typing import Any, Optional
 from uuid import UUID, uuid4
 
@@ -502,8 +503,8 @@ class Incident(Base):
     type: Mapped[str] = mapped_column(String(50), nullable=False)
     priority: Mapped[str] = mapped_column(String(20), nullable=False)
     location_address: Mapped[str | None] = mapped_column(Text, nullable=True)
-    location_lat: Mapped[float | None] = mapped_column(Numeric(10, 8), nullable=True)
-    location_lng: Mapped[float | None] = mapped_column(Numeric(11, 8), nullable=True)
+    location_lat: Mapped[Decimal | None] = mapped_column(Numeric(10, 8), nullable=True)
+    location_lng: Mapped[Decimal | None] = mapped_column(Numeric(11, 8), nullable=True)
     status: Mapped[str] = mapped_column(String(50), nullable=False, default="incoming")
     description: Mapped[str | None] = mapped_column(Text, nullable=True)
     contact: Mapped[str | None] = mapped_column(Text, nullable=True)  # Reporter/contact info (Melder/Anrufer)
